refactor: move node code to folder, remove unused code
